+++ src/Portfolio.js
@@ -5,6 +5,7 @@ import HeaderLogo from "./components/HeaderLogo";
 import { ProjectCard } from "./components/ProjectCard";
 import { ProjectList } from "./components/ProjectList";
 import { Resume } from "./components/Resume";
+import { TechSkillLevels } from "./components/TechSkillLevels";
 import { TechStack } from "./components/TechStack";
 
 export const Portfolio = () => {
@@ -30,34 +31,5 @@ export const Portfolio = () => {
       <Resume />
       <Footer />
     </div>
-  );
-};
-
-export const TechSkillLevels = ({ allTechObjects }) => {
-  return (
-    <section id="tech-skill-levels__block">
-      <ul className="tech__list list--no-padding-inline-left">
-        {allTechObjects.map((techObj) => (
-          <li className="tech__list-item list-item__no-bullet text__white-space--no-wrap" key={techObj.id}>
-            {" "}
-            <div className="tech-list-item__container">
-              <label htmlFor={`technology-${techObj.id}`}>
-                {techObj.name}:&nbsp;{" "}
-              </label>
-              <meter className="skill__meter"
-                id={`technology-${techObj.id}`}
-                max="5"
-                low="1"
-                high="5"
-                optimum="3"
-                value={techObj.skillLevel}
-              >
-                Skill Level &nbsp;{techObj.skillLevel}
-              </meter>
-            </div>
-          </li>
-        ))}
-      </ul>
-    </section>
   );
 };
